% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoSeed.r
\name{AutoSeed}
\alias{AutoSeed}
\title{Get disease-related genes from eDGAR, DrugBank and MalaCards}
\usage{
AutoSeed(search)
}
\arguments{
\item{search}{Name of the disease}
}
\value{
result$edgar: Containing Disease Name, OMIM ID and Genesymbol (Data comes from the eDGAR)
result$malacards: Containing genes related to the disease (Data comes from the MalaCards)
result$drugbank: Containing genes related to the disease (Data comes from the DrugBank)
}
\description{
Get disease-related genes from eDGAR, DrugBank and MalaCards
}
\examples{
result = AutoSeed("diabetes")
}
