\name{drugbank}
\alias{drugbank}
\docType{data}
\title{
Relationship between gene and disease
}
\description{
The data was collected in the drugbank. We got 27728 relationships between various disease and their related genes.
}
\usage{data("drugbank")}
\format{
  A data frame with 27728 observations on the following 2 variables.
}
\details{
There are two columns in the data. The first column is genes and the second column is diseases.
}
\source{
<https://www.drugbank.ca/>
}

\examples{
data(drugbank)
}
\keyword{datasets}
