% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoSeed.r
\name{drugbank_disease_gene}
\alias{drugbank_disease_gene}
\title{Get disease-related genes from DrugBank.}
\usage{
drugbank_disease_gene(search)
}
\arguments{
\item{search}{Name of the disease, character}
}
\value{
The genes related to the disease, list
}
\description{
Get disease-related genes from DrugBank.
}
\examples{
result = drugbank_disease_gene("diabetes")
}
