\name{edgar}
\alias{edgar}
\docType{data}
\title{
 Relationship between the gene and disease in edgar
}
\description{
The function "edgar_disease_gene()" will generate the relationship between gene and disease depend on this dataset.
}
\usage{data("edgar")}
\format{
  A data frame with 1038340 observations on the following 2 variables.
}
\details{
We got a total of 1038340 diseases and genes. There are two columns in the data. The first column is genes and the second column is diseases.
}
\source{
<http://edgar.biocomp.unibo.it/gene_disease_db/>
}

\examples{
data(edgar)
}
\keyword{datasets}
