% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoSeed.r
\name{edgar_disease_gene}
\alias{edgar_disease_gene}
\title{Get disease-related genes from eDGAR}
\usage{
edgar_disease_gene(disease)
}
\arguments{
\item{disease}{Name of the disease, character}
}
\value{
a vector containing genesymbol related to the disease
}
\description{
Get disease-related genes from eDGAR
}
\examples{
result = edgar_disease_gene("diabetes")
}
