\name{mala}
\alias{mala}
\docType{data}
\title{
The relationship between gene and disease
}
\description{
The data was collected in the Malacards. We got 241306 relationships between various disease and their related genes.
}
\usage{data("mala")}
\format{
  A data frame with 241306 observations on the following 2 variables.
}
\details{
We got a total of 241306 diseases and genes. There are two columns in the data. The first column is disease and the second column is gene.
}
\source{
<https://doi.org/10.1093/nar/gkw1012>
}

\examples{
data(mala)
}
\keyword{datasets}
