% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoSeed.r
\name{malacards_disease_gene}
\alias{malacards_disease_gene}
\title{Get disease-related genes from MalaCards}
\usage{
malacards_disease_gene(disease)
}
\arguments{
\item{disease}{Name of the disease}
}
\value{
The genes related to the disease, character vector
}
\description{
Get disease-related genes from MalaCards
}
\examples{
result = malacards_disease_gene("diabetes")
}
